+++ FITI-Trainer-iOS/FITI-Trainer-iOS/SignUp/TrainerCheck/TrainerCheckViewController.swift
@@ -1,29 +1,29 @@
 //
-//  FindPwViewController.swift
+//  TrainerCheckViewController.swift
 //  FITI-Trainer-iOS
 //
-//  Created by 박윤빈 on 2023/01/11.
+//  Created by 박윤빈 on 2023/01/10.
 //
 
 import Foundation
 import UIKit
 import SnapKit
 
-class FindPwViewController: UIViewController {
+class TrainerCheckViewController: UIViewController {
     
     var titleLabel : UILabel = {
         let label = UILabel()
         label.font = UIFont.boldSystemFont(ofSize: 25)
-        label.text = "비밀번호 찾기"
+        label.text = "트레이너 인증"
         label.textColor = UIColor.customColor(.blue)
         return label
     }()
     
-    private let emailTextField : UITextField = {
+    private let schoolTextField : UITextField = {
         let tf = UITextField()
         
         tf.attributedPlaceholder = NSAttributedString(
-                    string: "이메일을 입력해주세요.",
+                    string: "재학중인 학교를 입력해주세요",
                     attributes: [NSAttributedString.Key.foregroundColor: UIColor.customColor(.gray)]
                 )
         tf.layer.borderColor = UIColor.customColor(.gray).cgColor
@@ -36,9 +36,13 @@ class FindPwViewController: UIViewController {
         return tf
     }()
     
-    private let authTextField : UITextField = {
+    private let majorTextField : UITextField = {
         let tf = UITextField()
         tf.isEnabled = false
+        tf.attributedPlaceholder = NSAttributedString(
+                    string: "상세 전공을 입력해주세요",
+                    attributes: [NSAttributedString.Key.foregroundColor: UIColor.white]
+                )
         tf.layer.borderColor = UIColor.white.cgColor
         tf.layer.borderWidth = 2
         tf.layer.cornerRadius = 10
@@ -70,7 +74,7 @@ class FindPwViewController: UIViewController {
     let nextButton : UIButton = {
            let btn = UIButton()
             btn.backgroundColor = UIColor.customColor(.gray)
-            btn.setTitle("인증코드 받기", for: .normal)
+            btn.setTitle("다음", for: .normal)
             btn.setTitleColor(UIColor.white, for: .normal)
             btn.titleLabel?.font = UIFont(name: "Noto Sans", size: 0)
             btn.titleLabel?.font = UIFont.boldSystemFont(ofSize: 20)
@@ -102,8 +106,8 @@ class FindPwViewController: UIViewController {
         view.addSubview(titleLabel)
         view.addSubview(progressView)
         progressView.addSubview(grayView)
-        view.addSubview(emailTextField)
-        view.addSubview(authTextField)
+        view.addSubview(schoolTextField)
+        view.addSubview(majorTextField)
         view.addSubview(nextButton)
     }
     
@@ -124,15 +128,15 @@ class FindPwViewController: UIViewController {
             make.width.equalToSuperview().dividedBy(2)
         }
         
-        emailTextField.snp.makeConstraints { make in
+        schoolTextField.snp.makeConstraints { make in
             make.top.equalTo(progressView.snp.bottom).offset(68)
             make.leading.equalToSuperview().offset(20)
             make.trailing.equalToSuperview().offset(-20)
             make.height.equalTo(52)
         }
         
-        authTextField.snp.makeConstraints { make in
-            make.top.equalTo(emailTextField.snp.bottom).offset(20)
+        majorTextField.snp.makeConstraints { make in
+            make.top.equalTo(schoolTextField.snp.bottom).offset(20)
             make.leading.equalToSuperview().offset(20)
             make.trailing.equalToSuperview().offset(-20)
             make.height.equalTo(52)
@@ -155,49 +159,38 @@ class FindPwViewController: UIViewController {
     var nextBtn = 0
     
     @objc func handleSchoolTfDidChange(_ textField: UITextField) {
-        if(emailTextField.text != ""){
+        if(schoolTextField.text != ""){
             nextButton.backgroundColor = UIColor.customColor(.blue)
-            emailTextField.layer.borderColor = UIColor.customColor(.blue).cgColor
+            schoolTextField.layer.borderColor = UIColor.customColor(.blue).cgColor
         }
         nextBtn = 1
     }
     
     @objc func handleMajorTfDidChange(_ textField: UITextField) {
-        if(authTextField.text != ""){
+        if(majorTextField.text != ""){
             nextButton.backgroundColor = UIColor.customColor(.blue)
-            authTextField.layer.borderColor = UIColor.customColor(.blue).cgColor
+            majorTextField.layer.borderColor = UIColor.customColor(.blue).cgColor
             
         }
         nextBtn = 2
     }
     
     @objc func touchNextBtnEvent() {
-        
+        print(nextBtn)
         if(nextBtn == 2){
-            let nextVC = GradeTableViewController()
+            let nextVC = emailCheckViewController()
             navigationController?.pushViewController(nextVC, animated: true)
         } else if(nextBtn == 1){
             //major  textField 활성화
-            authTextField.isEnabled = true
-            authTextField.layer.borderColor = UIColor.customColor(.gray).cgColor
-            authTextField.attributedPlaceholder = NSAttributedString(
-                        string: "인증코드 6자리를 입력해주세요",
+            majorTextField.isEnabled = true
+            majorTextField.layer.borderColor = UIColor.customColor(.gray).cgColor
+            majorTextField.attributedPlaceholder = NSAttributedString(
+                        string: "상세 전공을 입력해주세요",
                         attributes: [NSAttributedString.Key.foregroundColor: UIColor.customColor(.gray)]
                     )
             //버튼 다시 회색으로
             nextButton.backgroundColor = UIColor.customColor(.gray)
-            nextButton.setTitle("다음", for: .normal)
     
-
-            self.grayView.snp.remakeConstraints({ make in
-                make.trailing.equalToSuperview()
-                make.height.equalTo(5)
-                make.width.equalToSuperview().dividedBy(2)
-            })
-                        
-            UIView.animate(withDuration: 0.5, animations: {
-                self.view.layoutIfNeeded()
-            }, completion: nil)
         }
     }
